+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Added
 
+* [HexaPDF::Layout::ColumnBox] for rendering content inside columns
 * [HexaPDF::Layout::MultiFrame] for placing boxes into  multiple frames
 * New configuration option 'debug' for enabling debug output